refactor: extract pure audio encoders and a shared Chord type
Split the WAV/MP3/PCM encoders out of audio_render.tsx into a
dependency-free audio_encoders module so they can be unit tested without
pulling in Tone or the offline render path, removing the need to widen
audio_render's public surface for tests. Introduce one canonical Chord
type in src/state/chord.ts and consume it from the store and the export
pipeline, replacing the shape that was duplicated inline.
